add jwt verification in delete api

+++ index.js
@@ -65,7 +65,6 @@ async function run() {
       if (email !== req.token_email) {
         return res.status(403).send({ message: "Forbidden access" });
       }
-
       const newFood = req.body;
       const result = await foodsCollection.insertOne(newFood);
       res.send(result);
@@ -115,9 +114,19 @@ async function run() {
       res.send(result);
     });
 
-    app.delete("/foods/:id", async (req, res) => {
+    app.delete("/foods/:id", verifyFirebaseToken, async (req, res) => {
       const id = req.params.id;
       const query = { _id: new ObjectId(id) };
+
+      const food = await foodsCollection.findOne(query);
+      if (!food) {
+        return res.status(404).send({ message: "Food not found." });
+      }
+
+      if (food.donator.email !== req.token_email) {
+        return res.status(403).send({ message: "Forbidden access" });
+      }
+
       const result = await foodsCollection.deleteOne(query);
       res.send(result);
     });
